+++ django_task/utils.py
@@ -86,6 +86,7 @@ def revoke_pending_tasks():
         n = queryset.count()
         print('revoking %s objects (%d) ...' % (model.__name__, n))
         #model.objects.all().delete()
+        import ipdb; ipdb.set_trace()
         queryset.update(status='REVOKED')
         counter += n