+++ example/example/management/commands/count_beans.py
@@ -7,11 +7,6 @@ class Command(TaskCommand):
         super(Command, self).add_arguments(parser)
         parser.add_argument('num_beans', type=int)
 
-    # def handle(self, *args, **options):
-    #     from example.models import CountBeansTask
-    #     from example.jobs import count_beans
-    #     self.run_job(CountBeansTask, count_beans, **options)
-
     def handle(self, *args, **options):
         from example.models import CountBeansTask
         self.run_task(CountBeansTask, **options)